# openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "Cliftin API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/programs": {
      "get": {
        "operationId": "listPrograms",
        "summary": "List programs",
        "tags": [
          "programs"
        ],
        "description": "Returns non-deleted programs. Use q for a case-insensitive name substring, one explicit sort value, and an opaque cursor for pagination.",
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false,
            "description": "Opaque cursor returned by a previous page."
          },
          {
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "example": "strength",
            "in": "query",
            "name": "q",
            "required": false,
            "description": "Case-insensitive substring match on program name."
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "name"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "-name"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "dateAdded"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "-dateAdded"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "sort",
            "required": false,
            "description": "Single sort field; prefix with - for descending order."
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "properties": {
                          "dateAdded": {
                            "anyOf": [
                              {
                                "format": "date-time",
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "id": {
                            "minimum": 1,
                            "type": "integer",
                            "example": 1
                          },
                          "isActive": {
                            "type": "boolean"
                          },
                          "isDeleted": {
                            "type": "boolean"
                          },
                          "isTemplate": {
                            "type": "boolean"
                          },
                          "name": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        },
                        "required": [
                          "dateAdded",
                          "id",
                          "isActive",
                          "isDeleted",
                          "isTemplate",
                          "name"
                        ]
                      }
                    },
                    "nextCursor": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "items"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          },
                          "path": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "message",
                          "path"
                        ]
                      }
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 400,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/programs/active": {
      "get": {
        "operationId": "getActiveProgram",
        "summary": "Get the active program",
        "tags": [
          "programs"
        ],
        "description": "Returns the program selected by Liftin, with legacy current-program fallback when no selection exists.",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "dateAdded": {
                      "anyOf": [
                        {
                          "format": "date-time",
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "id": {
                      "minimum": 1,
                      "type": "integer",
                      "example": 1
                    },
                    "isActive": {
                      "type": "boolean"
                    },
                    "isDeleted": {
                      "type": "boolean"
                    },
                    "isTemplate": {
                      "type": "boolean"
                    },
                    "name": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "dateAdded",
                    "id",
                    "isActive",
                    "isDeleted",
                    "isTemplate",
                    "name"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          },
                          "path": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "message",
                          "path"
                        ]
                      }
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 400,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/programs/active/next-routine": {
      "get": {
        "operationId": "getActiveProgramNextRoutine",
        "summary": "Get the active program next routine",
        "tags": [
          "programs"
        ],
        "description": "Returns the single up-next routine for the active program with normalized planned exercises and sets.",
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "lb"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "kg"
                  ]
                }
              ]
            },
            "example": "lb",
            "in": "query",
            "name": "unit",
            "required": false,
            "description": "Weight display unit. Defaults to lb where applicable."
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "allOf": [
                    {
                      "additionalProperties": false,
                      "type": "object",
                      "properties": {
                        "id": {
                          "examples": [
                            100
                          ],
                          "minimum": 1,
                          "type": "integer"
                        },
                        "isDeleted": {
                          "type": "boolean"
                        },
                        "isNext": {
                          "type": "boolean"
                        },
                        "name": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "program": {
                          "additionalProperties": false,
                          "type": "object",
                          "properties": {
                            "id": {
                              "examples": [
                                1
                              ],
                              "minimum": 1,
                              "type": "integer"
                            },
                            "name": {
                              "anyOf": [
                                {
                                  "type": "string"
                                },
                                {
                                  "type": "null"
                                }
                              ]
                            }
                          },
                          "required": [
                            "id",
                            "name"
                          ]
                        },
                        "week": {
                          "additionalProperties": false,
                          "type": "object",
                          "properties": {
                            "id": {
                              "examples": [
                                10
                              ],
                              "minimum": 1,
                              "type": "integer"
                            },
                            "number": {
                              "examples": [
                                1
                              ],
                              "minimum": 1,
                              "type": "integer"
                            }
                          },
                          "required": [
                            "id",
                            "number"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "isDeleted",
                        "isNext",
                        "name",
                        "program",
                        "week"
                      ]
                    },
                    {
                      "additionalProperties": false,
                      "type": "object",
                      "properties": {
                        "exercises": {
                          "type": "array",
                          "items": {
                            "additionalProperties": false,
                            "type": "object",
                            "properties": {
                              "exerciseId": {
                                "anyOf": [
                                  {
                                    "description": "Exercise definition primary key.",
                                    "minimum": 1,
                                    "type": "integer"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "id": {
                                "description": "Exercise configuration primary key.",
                                "minimum": 1,
                                "type": "integer"
                              },
                              "name": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "sets": {
                                "type": "array",
                                "items": {
                                  "additionalProperties": false,
                                  "type": "object",
                                  "properties": {
                                    "id": {
                                      "anyOf": [
                                        {
                                          "minimum": 1,
                                          "type": "integer"
                                        },
                                        {
                                          "type": "null"
                                        }
                                      ]
                                    },
                                    "reps": {
                                      "anyOf": [
                                        {
                                          "type": "number"
                                        },
                                        {
                                          "type": "null"
                                        }
                                      ]
                                    },
                                    "rpe": {
                                      "anyOf": [
                                        {
                                          "type": "number"
                                        },
                                        {
                                          "type": "null"
                                        }
                                      ]
                                    },
                                    "timeSeconds": {
                                      "anyOf": [
                                        {
                                          "type": "number"
                                        },
                                        {
                                          "type": "null"
                                        }
                                      ]
                                    },
                                    "weight": {
                                      "additionalProperties": false,
                                      "type": "object",
                                      "properties": {
                                        "unit": {
                                          "description": "Weight display unit. Defaults to lb where applicable.",
                                          "examples": [
                                            "lb"
                                          ],
                                          "anyOf": [
                                            {
                                              "type": "string",
                                              "enum": [
                                                "lb"
                                              ]
                                            },
                                            {
                                              "type": "string",
                                              "enum": [
                                                "kg"
                                              ]
                                            }
                                          ]
                                        },
                                        "value": {
                                          "anyOf": [
                                            {
                                              "type": "number"
                                            },
                                            {
                                              "type": "null"
                                            }
                                          ]
                                        }
                                      },
                                      "required": [
                                        "unit",
                                        "value"
                                      ]
                                    }
                                  },
                                  "required": [
                                    "id",
                                    "reps",
                                    "rpe",
                                    "timeSeconds",
                                    "weight"
                                  ]
                                }
                              }
                            },
                            "required": [
                              "exerciseId",
                              "id",
                              "name",
                              "sets"
                            ]
                          }
                        }
                      },
                      "required": [
                        "exercises"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          },
                          "path": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "message",
                          "path"
                        ]
                      }
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 400,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "status"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          },
                          "path": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "message",
                          "path"
                        ]
                      }
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 400,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/programs/{programId}": {
      "get": {
        "operationId": "getProgram",
        "summary": "Get program metadata",
        "tags": [
          "programs"
        ],
        "description": "Returns program metadata separately from its planned hierarchy. Direct lookup can return a soft-deleted program with isDeleted true.",
        "parameters": [
          {
            "schema": {
              "minimum": 1,
              "type": "integer"
            },
            "example": 1,
            "in": "path",
            "name": "programId",
            "required": true,
            "description": "Program primary key."
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "dateAdded": {
                      "anyOf": [
                        {
                          "format": "date-time",
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "id": {
                      "minimum": 1,
                      "type": "integer",
                      "example": 1
                    },
                    "isActive": {
                      "type": "boolean"
                    },
                    "isDeleted": {
                      "type": "boolean"
                    },
                    "isTemplate": {
                      "type": "boolean"
                    },
                    "name": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "dateAdded",
                    "id",
                    "isActive",
                    "isDeleted",
                    "isTemplate",
                    "name"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          },
                          "path": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "message",
                          "path"
                        ]
                      }
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 400,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": []
}